+++ packages/checkbox/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pikas-ui/checkbox
 
+## 1.0.16
+
+### Patch Changes
+
+- Create Textarea, Context Menu and Dropdown Menu
+- Updated dependencies
+  - @pikas-ui/icons@1.0.16
+  - @pikas-ui/styles@1.0.16
+  - @pikas-ui/text@1.0.16
+
 ## 1.0.15
 
 ### Patch Changes