+++ packages/loader/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/loader
 
+## 1.0.16
+
+### Patch Changes
+
+- Create Textarea, Context Menu and Dropdown Menu
+- Updated dependencies
+  - @pikas-ui/styles@1.0.16
+
 ## 1.0.15
 
 ### Patch Changes