+++ packages/table/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pikas-ui/table
 
+## 1.0.16
+
+### Patch Changes
+
+- Create Textarea, Context Menu and Dropdown Menu
+- Updated dependencies
+  - @pikas-ui/checkbox@1.0.16
+  - @pikas-ui/icons@1.0.16
+  - @pikas-ui/loader@1.0.16
+  - @pikas-ui/select@1.0.16
+  - @pikas-ui/styles@1.0.16
+
 ## 1.0.15
 
 ### Patch Changes